cloud_function/encrypting/requirements.txt
functions-framework==3.*
pandas
google-cloud-bigquery
loguru
pyarrow
PyCryptodome
rich